feat(pdf_extract): Add --draw-deleted flag
This commit introduces a new `--draw-deleted` flag to the `pdf_extract` tool.
This flag allows the user to control whether to draw bounding boxes for elements that have been marked as deleted during the classification process.

By default, deleted elements are not drawn. To include them in the output images, the `--draw-deleted` flag can be used.